# Address some predict/transform type instabilities