@ablaom 2024-04-24
Doesn't look like there's a significant difference. 

Before this PR:

```julia
julia> @time_imports import MLJBase
    413.2 ms  MLJBase 23.21% compilation time
```

After this PR:

```julia
@time_imports import MLJBase
    437.3 ms  MLJBase 22.21% compilation time
```